services/gradio_sql/requirements.txt
# requirements.txt
gradio==3.50.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0